move FolderSelect to component folder
